Ignore Serialization of Null values and solve Models property types issues

+++ commercetools.Sdk/commercetools.Sdk.Domain/ShippingMethods/CartScoreShippingRatePriceTier.cs
@@ -4,7 +4,7 @@
     public class CartScoreShippingRatePriceTier : ShippingRatePriceTier
     {
         // TODO See if fixed price and function can be combined in one class
-        public double Score { get; set; }
+        public long Score { get; set; }
         public Money Price { get; set; }
         public bool IsMatching { get; set; }
